Added Connection UI
Added Connection UI with a functioning buttons for Host, Client and Leave which are all contained in a canvas (putting the canvas made an event system automatically get added).
Also added an emptyobject renamed Network Manager with said script of the same name and a U net transport script too.